% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week1.R
\docType{data}
\name{data_week1}
\alias{data_week1}
\title{Week 1 Household Pulse Survey data}
\format{
\subsection{\code{data_week1}}{

A data frame with 74,413 rows and 82 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week1
}
\description{
Week 1 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
