% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week10.R
\docType{data}
\name{data_week10}
\alias{data_week10}
\title{Week 10 Household Pulse Survey data}
\format{
\subsection{\code{data_week10}}{

A data frame with 90,767 rows and 105 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week10
}
\description{
Week 10 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
