% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week11.R
\docType{data}
\name{data_week11}
\alias{data_week11}
\title{Week 11 Household Pulse Survey data}
\format{
\subsection{\code{data_week11}}{

A data frame with 91,605 rows and 105 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week11
}
\description{
Week 11 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
