% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week12.R
\docType{data}
\name{data_week12}
\alias{data_week12}
\title{Week 12 Household Pulse Survey data}
\format{
\subsection{\code{data_week12}}{

A data frame with 86,792 rows and 105 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week12
}
\description{
Week 12 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
