% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week13.R
\docType{data}
\name{data_week13}
\alias{data_week13}
\title{Week 13 Household Pulse Survey data}
\format{
\subsection{\code{data_week13}}{

A data frame with 109,051 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week13
}
\description{
Week 13 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
