% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week15.R
\docType{data}
\name{data_week15}
\alias{data_week15}
\title{Week 15 Household Pulse Survey data}
\format{
\subsection{\code{data_week15}}{

A data frame with 99,302 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week15
}
\description{
Week 15 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
