% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week16.R
\docType{data}
\name{data_week16}
\alias{data_week16}
\title{Week 16 Household Pulse Survey data}
\format{
\subsection{\code{data_week16}}{

A data frame with 95,604 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week16
}
\description{
Week 16 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
