% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week17.R
\docType{data}
\name{data_week17}
\alias{data_week17}
\title{Week 17 Household Pulse Survey data}
\format{
\subsection{\code{data_week17}}{

A data frame with 88,716 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week17
}
\description{
Week 17 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
