% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week18.R
\docType{data}
\name{data_week18}
\alias{data_week18}
\title{Week 18 Household Pulse Survey data}
\format{
\subsection{\code{data_week18}}{

A data frame with 58,729 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week18
}
\description{
Week 18 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
