% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week19.R
\docType{data}
\name{data_week19}
\alias{data_week19}
\title{Week 19 Household Pulse Survey data}
\format{
\subsection{\code{data_week19}}{

A data frame with 71,939 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week19
}
\description{
Week 19 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
