% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week2.R
\docType{data}
\name{data_week2}
\alias{data_week2}
\title{Week 2 Household Pulse Survey data}
\format{
\subsection{\code{data_week2}}{

A data frame with 41,996 rows and 82 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week2
}
\description{
Week 2 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
