% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week20.R
\docType{data}
\name{data_week20}
\alias{data_week20}
\title{Week 20 Household Pulse Survey data}
\format{
\subsection{\code{data_week20}}{

A data frame with 72,484 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week20
}
\description{
Week 20 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
