% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week21.R
\docType{data}
\name{data_week21}
\alias{data_week21}
\title{Week 21 Household Pulse Survey data}
\format{
\subsection{\code{data_week21}}{

A data frame with 69,944 rows and 188 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week21
}
\description{
Week 21 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
