% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week22.R
\docType{data}
\name{data_week22}
\alias{data_week22}
\title{Week 22 Household Pulse Survey data}
\format{
\subsection{\code{data_week22}}{

A data frame with 68,348 rows and 204 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week22
}
\description{
Week 22 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
