% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week23.R
\docType{data}
\name{data_week23}
\alias{data_week23}
\title{Week 23 Household Pulse Survey data}
\format{
\subsection{\code{data_week23}}{

A data frame with 80,567 rows and 204 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week23
}
\description{
Week 23 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
