% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week24.R
\docType{data}
\name{data_week24}
\alias{data_week24}
\title{Week 24 Household Pulse Survey data}
\format{
\subsection{\code{data_week24}}{

A data frame with 77,122 rows and 204 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week24
}
\description{
Week 24 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
