% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week25.R
\docType{data}
\name{data_week25}
\alias{data_week25}
\title{Week 25 Household Pulse Survey data}
\format{
\subsection{\code{data_week25}}{

A data frame with 77,788 rows and 204 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week25
}
\description{
Week 25 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
