% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week26.R
\docType{data}
\name{data_week26}
\alias{data_week26}
\title{Week 26 Household Pulse Survey data}
\format{
\subsection{\code{data_week26}}{

A data frame with 78,306 rows and 204 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week26
}
\description{
Week 26 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
