% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week27.R
\docType{data}
\name{data_week27}
\alias{data_week27}
\title{Week 27 Household Pulse Survey data}
\format{
\subsection{\code{data_week27}}{

A data frame with 77,104 rows and 204 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week27
}
\description{
Week 27 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
