% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week28.R
\docType{data}
\name{data_week28}
\alias{data_week28}
\title{Week 28 Household Pulse Survey data}
\format{
\subsection{\code{data_week28}}{

A data frame with 68,913 rows and 239 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week28
}
\description{
Week 28 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
