% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week29.R
\docType{data}
\name{data_week29}
\alias{data_week29}
\title{Week 29 Household Pulse Survey data}
\format{
\subsection{\code{data_week29}}{

A data frame with 78,467 rows and 239 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week29
}
\description{
Week 29 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
