% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week3.R
\docType{data}
\name{data_week3}
\alias{data_week3}
\title{Week 3 Household Pulse Survey data}
\format{
\subsection{\code{data_week3}}{

A data frame with 132,961 rows and 82 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week3
}
\description{
Week 3 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
