% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week30.R
\docType{data}
\name{data_week30}
\alias{data_week30}
\title{Week 30 Household Pulse Survey data}
\format{
\subsection{\code{data_week30}}{

A data frame with 72,897 rows and 239 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week30
}
\description{
Week 30 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
