% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week31.R
\docType{data}
\name{data_week31}
\alias{data_week31}
\title{Week 31 Household Pulse Survey data}
\format{
\subsection{\code{data_week31}}{

A data frame with 70,854 rows and 239 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week31
}
\description{
Week 31 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
