% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week32.R
\docType{data}
\name{data_week32}
\alias{data_week32}
\title{Week 32 Household Pulse Survey data}
\format{
\subsection{\code{data_week32}}{

A data frame with 68,067 rows and 239 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week32
}
\description{
Week 32 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
