% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week33.R
\docType{data}
\name{data_week33}
\alias{data_week33}
\title{Week 33 Household Pulse Survey data}
\format{
\subsection{\code{data_week33}}{

A data frame with 66,262 rows and 239 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week33
}
\description{
Week 33 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
