% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week34.R
\docType{data}
\name{data_week34}
\alias{data_week34}
\title{Week 34 Household Pulse Survey data}
\format{
\subsection{\code{data_week34}}{

A data frame with 64,562 rows and 202 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week34
}
\description{
Week 34 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
