% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week35.R
\docType{data}
\name{data_week35}
\alias{data_week35}
\title{Week 35 Household Pulse Survey data}
\format{
\subsection{\code{data_week35}}{

A data frame with 68,799 rows and 202 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week35
}
\description{
Week 35 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
