% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week36.R
\docType{data}
\name{data_week36}
\alias{data_week36}
\title{Week 36 Household Pulse Survey data}
\format{
\subsection{\code{data_week36}}{

A data frame with 69,114 rows and 202 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week36
}
\description{
Week 36 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
