% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week37.R
\docType{data}
\name{data_week37}
\alias{data_week37}
\title{Week 37 Household Pulse Survey data}
\format{
\subsection{\code{data_week37}}{

A data frame with 63,536 rows and 202 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week37
}
\description{
Week 37 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
