% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week38.R
\docType{data}
\name{data_week38}
\alias{data_week38}
\title{Week 38 Household Pulse Survey data}
\format{
\subsection{\code{data_week38}}{

A data frame with 59,833 rows and 202 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week38
}
\description{
Week 38 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
