% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week39.R
\docType{data}
\name{data_week39}
\alias{data_week39}
\title{Week 39 Household Pulse Survey data}
\format{
\subsection{\code{data_week39}}{

A data frame with 57,064 rows and 202 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week39
}
\description{
Week 39 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
