% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week4.R
\docType{data}
\name{data_week4}
\alias{data_week4}
\title{Week 4 Household Pulse Survey data}
\format{
\subsection{\code{data_week4}}{

A data frame with 101,215 rows and 82 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week4
}
\description{
Week 4 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
