% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week40.R
\docType{data}
\name{data_week40}
\alias{data_week40}
\title{Week 40 Household Pulse Survey data}
\format{
\subsection{\code{data_week40}}{

A data frame with 60,826 rows and 208 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week40
}
\description{
Week 40 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
