% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week41.R
\docType{data}
\name{data_week41}
\alias{data_week41}
\title{Week 41 Household Pulse Survey data}
\format{
\subsection{\code{data_week41}}{

A data frame with 74,995 rows and 208 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week41
}
\description{
Week 41 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
