% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week42.R
\docType{data}
\name{data_week42}
\alias{data_week42}
\title{Week 42 Household Pulse Survey data}
\format{
\subsection{\code{data_week42}}{

A data frame with 75,482 rows and 208 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week42
}
\description{
Week 42 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
