% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week43.R
\docType{data}
\name{data_week43}
\alias{data_week43}
\title{Week 43 Household Pulse Survey data}
\format{
\subsection{\code{data_week43}}{

A data frame with 84,158 rows and 219 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week43
}
\description{
Week 43 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
