% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week44.R
\docType{data}
\name{data_week44}
\alias{data_week44}
\title{Week 44 Household Pulse Survey data}
\format{
\subsection{\code{data_week44}}{

A data frame with 63,769 rows and 219 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week44
}
\description{
Week 44 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
