% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week45.R
\docType{data}
\name{data_week45}
\alias{data_week45}
\title{Week 45 Household Pulse Survey data}
\format{
\subsection{\code{data_week45}}{

A data frame with 61,767 rows and 219 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week45
}
\description{
Week 45 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
