% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week46.R
\docType{data}
\name{data_week46}
\alias{data_week46}
\title{Week 46 Household Pulse Survey data}
\format{
\subsection{\code{data_week46}}{

A data frame with 62,826 rows and 240 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week46
}
\description{
Week 46 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
