% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week47.R
\docType{data}
\name{data_week47}
\alias{data_week47}
\title{Week 47 Household Pulse Survey data}
\format{
\subsection{\code{data_week47}}{

A data frame with 58,304 rows and 240 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week47
}
\description{
Week 47 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
