% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week48.R
\docType{data}
\name{data_week48}
\alias{data_week48}
\title{Week 48 Household Pulse Survey data}
\format{
\subsection{\code{data_week48}}{

A data frame with 46,801 rows and 240 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week48
}
\description{
Week 48 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
