% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week49.R
\docType{data}
\name{data_week49}
\alias{data_week49}
\title{Week 49 Household Pulse Survey data}
\format{
\subsection{\code{data_week49}}{

A data frame with 50,937 rows and 244 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week49
}
\description{
Week 49 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
