% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week5.R
\docType{data}
\name{data_week5}
\alias{data_week5}
\title{Week 5 Household Pulse Survey data}
\format{
\subsection{\code{data_week5}}{

A data frame with 105,066 rows and 82 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week5
}
\description{
Week 5 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
