% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week50.R
\docType{data}
\name{data_week50}
\alias{data_week50}
\title{Week 50 Household Pulse Survey data}
\format{
\subsection{\code{data_week50}}{

A data frame with 42,040 rows and 244 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week50
}
\description{
Week 50 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
