% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week52.R
\docType{data}
\name{data_week52}
\alias{data_week52}
\title{Week 52 Household Pulse Survey data}
\format{
\subsection{\code{data_week52}}{

A data frame with 70,685 rows and 243 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week52
}
\description{
Week 52 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
