% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week53.R
\docType{data}
\name{data_week53}
\alias{data_week53}
\title{Week 53 Household Pulse Survey data}
\format{
\subsection{\code{data_week53}}{

A data frame with 68,504 rows and 243 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week53
}
\description{
Week 53 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
