% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week54.R
\docType{data}
\name{data_week54}
\alias{data_week54}
\title{Week 54 Household Pulse Survey data}
\format{
\subsection{\code{data_week54}}{

A data frame with 75,709 rows and 243 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week54
}
\description{
Week 54 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
