% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week55.R
\docType{data}
\name{data_week55}
\alias{data_week55}
\title{Week 55 Household Pulse Survey data}
\format{
\subsection{\code{data_week55}}{

A data frame with 72,738 rows and 244 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week55
}
\description{
Week 55 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
