% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week56.R
\docType{data}
\name{data_week56}
\alias{data_week56}
\title{Week 56 Household Pulse Survey data}
\format{
\subsection{\code{data_week56}}{

A data frame with 61,927 rows and 244 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week56
}
\description{
Week 56 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
