% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week57.R
\docType{data}
\name{data_week57}
\alias{data_week57}
\title{Week 57 Household Pulse Survey data}
\format{
\subsection{\code{data_week57}}{

A data frame with 59,290 rows and 244 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week57
}
\description{
Week 57 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
