% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week58.R
\docType{data}
\name{data_week58}
\alias{data_week58}
\title{Week 58 Household Pulse Survey data}
\format{
\subsection{\code{data_week58}}{

A data frame with 64,792 rows and 221 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week58
}
\description{
Week 58 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
