% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week59.R
\docType{data}
\name{data_week59}
\alias{data_week59}
\title{Week 59 Household Pulse Survey data}
\format{
\subsection{\code{data_week59}}{

A data frame with 63,802 rows and 221 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week59
}
\description{
Week 59 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
