% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week6.R
\docType{data}
\name{data_week6}
\alias{data_week6}
\title{Week 6 Household Pulse Survey data}
\format{
\subsection{\code{data_week6}}{

A data frame with 83,302 rows and 84 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week6
}
\description{
Week 6 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
