% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week60.R
\docType{data}
\name{data_week60}
\alias{data_week60}
\title{Week 60 Household Pulse Survey data}
\format{
\subsection{\code{data_week60}}{

A data frame with 68,830 rows and 221 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week60
}
\description{
Week 60 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
