% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week61.R
\docType{data}
\name{data_week61}
\alias{data_week61}
\title{Week 61 Household Pulse Survey data}
\format{
\subsection{\code{data_week61}}{

A data frame with 68,454 rows and 247 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week61
}
\description{
Week 61 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
