% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week62.R
\docType{data}
\name{data_week62}
\alias{data_week62}
\title{Week 62 Household Pulse Survey data}
\format{
\subsection{\code{data_week62}}{

A data frame with 72,839 rows and 247 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week62
}
\description{
Week 62 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
