% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week63.R
\docType{data}
\name{data_week63}
\alias{data_week63}
\title{Week 63 Household Pulse Survey data}
\format{
\subsection{\code{data_week63}}{

A data frame with 79,371 rows and 247 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week63
}
\description{
Week 63 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
