% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week7.R
\docType{data}
\name{data_week7}
\alias{data_week7}
\title{Week 7 Household Pulse Survey data}
\format{
\subsection{\code{data_week7}}{

A data frame with 73,472 rows and 105 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week7
}
\description{
Week 7 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
