% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week8.R
\docType{data}
\name{data_week8}
\alias{data_week8}
\title{Week 8 Household Pulse Survey data}
\format{
\subsection{\code{data_week8}}{

A data frame with 108,062 rows and 105 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week8
}
\description{
Week 8 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
