% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_week9.R
\docType{data}
\name{data_week9}
\alias{data_week9}
\title{Week 9 Household Pulse Survey data}
\format{
\subsection{\code{data_week9}}{

A data frame with 98,663 rows and 105 columns:
\describe{
\item{SCRAM}{Response ID}
\item{WEEK}{Week Number}
...
}
}
}
\source{
\url{https://github.com/prasadbhoite/HPS_PUF/raw/main/Data}
}
\usage{
data_week9
}
\description{
Week 9 data from the US census's Household Pulse Survey.
}
\keyword{datasets}
